requirements.txt
#requirements.txt
requests
icalendar
wyze-sdk
python-dotenv
schedule
argparse
pytz